Minor Changes to v0.2
CHANGE LOG:
1. Changed variable "timer" both in the implementation file
and the header file from unsigned short to short
2. In the implementation file I changed the delay from 1 second
  to 4 ms (1 second is a bit too long!)
PROBLEMS:
I realized why the test driver wouldn't succeed. The timing is tricky
since my test program required an event to happen exactly at t=30s
(which is common sense for an event that happens only once) However,
the delay cycle might not be in phase of the time the event is happening.